requirements.txt
py-clob-client==1.0.10
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.1
sqlmodel==0.0.14
rich==13.7.0
pydantic==2.5.2
pydantic-settings==2.1.0
pandas==2.1.3
openai==1.3.7  # For Grok API (OpenAI-compatible)
eth-account==0.10.0
web3==6.11.3
apscheduler==3.10.4  # For task scheduling
loguru==0.7.2  # Better logging
